refactor: extract ButtonIconBase
